src/main/java/ch/hsr/xclavis/crypto/HMACSHA.java
/*
 * Copyright (c) 2015, Gian Poltéra
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are met:
 *
 * 1.	Redistributions of source code must retain the above copyright notice,
 *   	this list of conditions and the following disclaimer.
 * 2.	Redistributions in binary form must reproduce the above copyright 
 *   	notice, this list of conditions and the following disclaimer in the 
 *   	documentation and/or other materials provided with the distribution.
 * 3.	Neither the name of HSR University of Applied Sciences Rapperswil nor 
 * 	the names of its contributors may be used to endorse or promote products
 * 	derived from this software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
 * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */
package ch.hsr.xclavis.crypto;

import org.bouncycastle.crypto.Digest;
import org.bouncycastle.crypto.digests.SHA256Digest;
import org.bouncycastle.crypto.digests.SHA512Digest;
import org.bouncycastle.crypto.macs.HMac;
import org.bouncycastle.crypto.params.KeyParameter;

/**
 *
 * @author Gian
 */
public class HMACSHA {
    private final static String KEY = "XClavis Key Exchange";
    private byte[] inputBuffer, outputBuffer, derivatedKey;
    private Digest digest;

    public HMACSHA(int length) {
        this.outputBuffer = new byte[length / Byte.SIZE];
        this.inputBuffer = KEY.getBytes();
        
        if (length == 256) {
            digest = new SHA256Digest();
        } else if (length == 512) {
            digest = new SHA512Digest();
        }        
    }

    public byte[] getDerivatedKey(int length, byte[] key) {
        derivatedKey = new byte[length / Byte.SIZE];

        HMac mac = new HMac(digest);
        mac.init(new KeyParameter(key));
        mac.update(inputBuffer, 0, inputBuffer.length);
        mac.doFinal(outputBuffer, 0);

        // Source Array, From Source, Destination Array, To Destination, Count
        System.arraycopy(outputBuffer, 0, derivatedKey, 0, derivatedKey.length);

        return derivatedKey;
    }
}
